requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
scipy>=1.10
scikit-learn>=1.3
statsmodels>=0.14
anthropic>=0.30
streamlit>=1.35
python-dotenv>=1.0
